// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  PRESIDENT: 'PRESIDENT',
  COORDINATOR: 'COORDINATOR',
  ACCOUNTANT: 'ACCOUNTANT',
  MEMBER: 'MEMBER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const SavingStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type SavingStatus = (typeof SavingStatus)[keyof typeof SavingStatus]


export const TransactionType = {
  DEPOSIT: 'DEPOSIT',
  WITHDRAW: 'WITHDRAW',
  ADJUSTMENT: 'ADJUSTMENT'
} as const

export type TransactionType = (typeof TransactionType)[keyof typeof TransactionType]


export const ReminderStatus = {
  PENDING: 'PENDING',
  SENT: 'SENT',
  FAILED: 'FAILED'
} as const

export type ReminderStatus = (typeof ReminderStatus)[keyof typeof ReminderStatus]


export const ExpenseStatus = {
  ACTIVE: 'ACTIVE',
  CANCELLED: 'CANCELLED'
} as const

export type ExpenseStatus = (typeof ExpenseStatus)[keyof typeof ExpenseStatus]


export const LedgerType = {
  SAVING: 'SAVING',
  EXPENSE: 'EXPENSE',
  ADJUSTMENT: 'ADJUSTMENT'
} as const

export type LedgerType = (typeof LedgerType)[keyof typeof LedgerType]